+++ Bencodex/Types/Dictionary.cs
@@ -15,8 +15,10 @@ namespace Bencodex.Types
     /// <summary>
     /// Represents Bencodex dictionaries.
     /// </summary>
+    [Pure]
     public class Dictionary :
         IValue,
+        IEquatable<Dictionary>,
         IEquatable<IImmutableDictionary<IKey, IValue>>,
         IImmutableDictionary<IKey, IValue>
     {
@@ -36,14 +38,7 @@ namespace Bencodex.Types
 
         private static readonly byte[] _unicodeKeyPrefix = new byte[1] { 0x75 };  // 'u'
 
-        private static IComparer<ValueTuple<byte?, byte[]>> keyPairComparer =
-            new CompositeComparer<byte?, byte[]>(
-                Comparer<byte?>.Default,
-                default(ByteArrayComparer)
-            );
-
-        private ImmutableDictionary<IKey, IValue>? _dict;
-        private KeyValuePair<IKey, IValue>[]? _pairs;
+        private ImmutableSortedDictionary<IKey, IValue> _dict;
         private ImmutableArray<byte>? _hash;
         private long _encodingLength = -1;
 
@@ -53,60 +48,39 @@ namespace Bencodex.Types
         /// <param name="pairs">Key-value pairs to include.  If there are duplicated keys,
         /// later pairs overwrite earlier ones.</param>
         public Dictionary(IEnumerable<KeyValuePair<IKey, IValue>> pairs)
+            : this(pairs.ToImmutableSortedDictionary(KeyComparer.Instance))
         {
-            _pairs = pairs.ToArray();
         }
 
         /// <summary>
         /// Creates a <see cref="Dictionary"/> instance with the <paramref name="content"/>.
         /// </summary>
         /// <param name="content">The dictionary content.</param>
-        public Dictionary(in ImmutableDictionary<IKey, IValue> content)
+        public Dictionary(in ImmutableSortedDictionary<IKey, IValue> content)
         {
-            _dict = content;
+            _dict = content.KeyComparer == KeyComparer.Instance
+                ? content
+                : content.WithComparers(KeyComparer.Instance);
         }
 
-        public int Count => Value.Count;
+        /// <inheritdoc cref="IReadOnlyCollection{T}.Count"/>
+        public int Count => _dict.Count;
 
-        public IEnumerable<IKey> Keys =>
-            Value.Keys;
+        /// <inheritdoc cref="IReadOnlyDictionary{TKey,TValue}.Keys"/>
+        public IEnumerable<IKey> Keys => _dict.Keys;
 
-        public IEnumerable<IValue> Values =>
-            Value.Values;
+        /// <inheritdoc cref="IReadOnlyDictionary{TKey,TValue}.Values"/>
+        public IEnumerable<IValue> Values => _dict.Values;
 
         /// <inheritdoc cref="IValue.Type"/>
-        [Pure]
         public ValueType Type => ValueType.Dictionary;
 
         /// <inheritdoc cref="IValue.Fingerprint"/>
-        [Pure]
         public Fingerprint Fingerprint
         {
             get
             {
-                if (_dict is { } d && d.IsEmpty)
-                {
-                    return EmptyFingerprint;
-                }
-                else if (_pairs is { } p && p.Length < 1)
-                {
-                    return EmptyFingerprint;
-                }
-
-                var pairs = _dict ?? _pairs ?? Enumerable.Empty<KeyValuePair<IKey, IValue>>();
-                var fDict = new SortedDictionary<Fingerprint, Fingerprint>(
-                    new FingerprintComparer()
-                );
-                foreach (KeyValuePair<IKey, IValue> kv in pairs)
-                {
-                    Fingerprint keyF = kv.Key.Fingerprint;
-                    if (!fDict.ContainsKey(keyF))
-                    {
-                        fDict[keyF] = kv.Value.Fingerprint;
-                    }
-                }
-
-                if (!fDict.Any())
+                if (_dict.Count < 1)
                 {
                     return EmptyFingerprint;
                 }
@@ -116,11 +90,11 @@ namespace Bencodex.Types
                     long encLength = 2L;
                     SHA1 sha1 = SHA1.Create();
                     sha1.Initialize();
-                    foreach (KeyValuePair<Fingerprint, Fingerprint> pair in fDict)
+                    foreach (KeyValuePair<IKey, IValue> pair in this)
                     {
-                        byte[] fp = pair.Key.Serialize();
+                        byte[] fp = pair.Key.Fingerprint.Serialize();
                         sha1.TransformBlock(fp, 0, fp.Length, null, 0);
-                        fp = pair.Value.Serialize();
+                        fp = pair.Value.Fingerprint.Serialize();
                         sha1.TransformBlock(fp, 0, fp.Length, null, 0);
                         encLength += pair.Key.EncodingLength +
                                      pair.Value.EncodingLength;
@@ -140,25 +114,19 @@ namespace Bencodex.Types
         }
 
         /// <inheritdoc cref="IValue.EncodingLength"/>
-        [Pure]
         public long EncodingLength =>
             _encodingLength >= 0
                 ? _encodingLength
                 : _encodingLength = _dictionaryPrefix.LongLength
-                    + Value.Sum(kv => kv.Key.EncodingLength + kv.Value.EncodingLength)
+                    + _dict.Sum(kv => kv.Key.EncodingLength + kv.Value.EncodingLength)
                     + CommonVariables.Suffix.LongLength;
 
         /// <inheritdoc cref="IValue.Inspection"/>
-        [Pure]
         public string Inspection
         {
             get
             {
-                if (_dict is { } d && d.IsEmpty)
-                {
-                    return "{}";
-                }
-                else if (_pairs is { } p && p.Length < 1)
+                if (_dict.Count < 1)
                 {
                     return "{}";
                 }
@@ -171,25 +139,8 @@ namespace Bencodex.Types
             }
         }
 
-        [Pure]
-        private ImmutableDictionary<IKey, IValue> Value
-        {
-            get
-            {
-                if (!(_dict is { } dict))
-                {
-                    dict = _pairs.ToImmutableDictionary();
-                    _dict = dict;
-                    _pairs = null;
-                }
-
-                return dict;
-            }
-        }
-
         /// <inheritdoc cref="IReadOnlyDictionary{TKey,TValue}.this[TKey]"/>
-        [Pure]
-        public IValue this[IKey key] => Value[key];
+        public IValue this[IKey key] => _dict[key];
 
         public IValue this[string key] => this[(IKey)new Text(key)];
 
@@ -197,15 +148,14 @@ namespace Bencodex.Types
 
         public IValue this[byte[] key] => this[(IKey)new Binary(key)];
 
-        public IEnumerator<KeyValuePair<IKey, IValue>> GetEnumerator() => Value?.GetEnumerator()
-            ?? Enumerable.Empty<KeyValuePair<IKey, IValue>>().GetEnumerator();
+        /// <inheritdoc cref="IEnumerable{T}.GetEnumerator()"/>
+        public IEnumerator<KeyValuePair<IKey, IValue>> GetEnumerator() => _dict.GetEnumerator();
 
-        IEnumerator IEnumerable.GetEnumerator() =>
-            (IEnumerator<KeyValuePair<IKey, IValue>>)GetEnumerator();
+        /// <inheritdoc cref="IEnumerable.GetEnumerator()"/>
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
         /// <inheritdoc cref="IReadOnlyDictionary{TKey,TValue}.ContainsKey(TKey)"/>
-        [Pure]
-        public bool ContainsKey(IKey key) => Value.ContainsKey(key);
+        public bool ContainsKey(IKey key) => _dict.ContainsKey(key);
 
         public bool ContainsKey(string key) => ContainsKey((IKey)new Text(key));
 
@@ -214,12 +164,12 @@ namespace Bencodex.Types
         public bool ContainsKey(byte[] key) => ContainsKey((IKey)new Binary(key));
 
         /// <inheritdoc cref="IReadOnlyDictionary{TKey,TValue}.TryGetValue(TKey, out TValue)"/>
-        [Pure]
         public bool TryGetValue(IKey key, out IValue value) =>
-            Value.TryGetValue(key, out value);
+            _dict.TryGetValue(key, out value);
 
+        /// <inheritdoc cref="IImmutableDictionary{TKey,TValue}.Add(TKey, TValue)"/>
         public IImmutableDictionary<IKey, IValue> Add(IKey key, IValue value) =>
-            new Dictionary(Value.Add(key, value));
+            new Dictionary(_dict.Add(key, value));
 
         public Dictionary Add(string key, IValue value) =>
             (Dictionary)Add((IKey)new Text(key), value);
@@ -277,32 +227,30 @@ namespace Bencodex.Types
         public Dictionary Add(byte[] key, IEnumerable<IValue> value) =>
             Add(key, (IValue)new List(value));
 
+        /// <inheritdoc cref="IImmutableDictionary{TKey,TValue}.AddRange"/>
         public IImmutableDictionary<IKey, IValue> AddRange(
             IEnumerable<KeyValuePair<IKey, IValue>> pairs
         ) =>
-            new Dictionary(Value.AddRange(pairs));
+            new Dictionary(_dict.AddRange(pairs));
 
         /// <inheritdoc cref="IImmutableDictionary{TKey,TValue}.Clear()"/>
-        [Pure]
         public IImmutableDictionary<IKey, IValue> Clear() => Empty;
 
         /// <inheritdoc
         /// cref="IImmutableDictionary{TKey,TValue}.Contains(KeyValuePair{TKey, TValue})"/>
-        [Pure]
-        public bool Contains(KeyValuePair<IKey, IValue> pair) => Value.Contains(pair);
+        public bool Contains(KeyValuePair<IKey, IValue> pair) => _dict.Contains(pair);
 
         /// <inheritdoc cref="IImmutableDictionary{TKey,TValue}.Remove(TKey)"/>
-        [Pure]
         public IImmutableDictionary<IKey, IValue> Remove(IKey key) =>
-            Value.IsEmpty ? this : new Dictionary(Value.Remove(key));
+            _dict.Count < 1 ? this : new Dictionary(_dict.Remove(key));
 
         /// <inheritdoc cref="IImmutableDictionary{TKey,TValue}.RemoveRange(IEnumerable{TKey})"/>
-        [Pure]
         public IImmutableDictionary<IKey, IValue> RemoveRange(IEnumerable<IKey> keys) =>
-            Value.IsEmpty ? this : new Dictionary(Value.RemoveRange(keys));
+            _dict.Count < 1 ? this : new Dictionary(_dict.RemoveRange(keys));
 
+        /// <inheritdoc cref="IImmutableDictionary{TKey,TValue}.SetItem(TKey,TValue)"/>
         public IImmutableDictionary<IKey, IValue> SetItem(IKey key, IValue value) =>
-            new Dictionary(Value.SetItem(key, value));
+            new Dictionary(_dict.SetItem(key, value));
 
         public Dictionary SetItem(IKey key, string value) =>
             (Dictionary)SetItem(key, (IValue)new Text(value));
@@ -397,15 +345,15 @@ namespace Bencodex.Types
         public Dictionary SetItem(byte[] key, IEnumerable<IValue> value) =>
             SetItem(key, (IValue)new List(value));
 
+        /// <inheritdoc cref="IImmutableDictionary{TKey,TValue}.SetItems"/>
         public IImmutableDictionary<IKey, IValue> SetItems(
             IEnumerable<KeyValuePair<IKey, IValue>> items
         ) =>
-            new Dictionary(Value.SetItems(items));
+            new Dictionary(_dict.SetItems(items));
 
         /// <inheritdoc cref="IImmutableDictionary{TKey,TValue}.TryGetKey(TKey, out TKey)"/>
-        [Pure]
         public bool TryGetKey(IKey equalKey, out IKey actualKey) =>
-            Value.TryGetKey(equalKey, out actualKey);
+            _dict.TryGetKey(equalKey, out actualKey);
 
         public T GetValue<T>(string name)
             where T : IValue
@@ -433,12 +381,26 @@ namespace Bencodex.Types
             };
 
         /// <inheritdoc cref="IEquatable{T}.Equals(T)"/>
-        [Pure]
+        public bool Equals(Dictionary other)
+        {
+            if (Count != other.Count)
+            {
+                return false;
+            }
+            else if (_hash is { } && other._hash is { })
+            {
+                return Fingerprint.Equals(other.Fingerprint);
+            }
+
+            return ((IEquatable<IImmutableDictionary<IKey, IValue>>)this).Equals(other);
+        }
+
+        /// <inheritdoc cref="IEquatable{T}.Equals(T)"/>
         bool IEquatable<IImmutableDictionary<IKey, IValue>>.Equals(
             IImmutableDictionary<IKey, IValue> other
         )
         {
-            if (Value.LongCount() != other.LongCount())
+            if (_dict.Count != other.Count)
             {
                 return false;
             }
@@ -456,54 +418,29 @@ namespace Bencodex.Types
         }
 
         /// <inheritdoc cref="IEquatable{T}.Equals(T)"/>
-        [Pure]
         bool IEquatable<IValue>.Equals(IValue other) =>
             other is Dictionary o &&
             ((IEquatable<IImmutableDictionary<IKey, IValue>>)this).Equals(o);
 
         /// <inheritdoc cref="object.GetHashCode()"/>
-        [Pure]
-        public override int GetHashCode() => Value.GetHashCode();
+        public override int GetHashCode() => _dict.GetHashCode();
 
         /// <inheritdoc cref="IValue.EncodeIntoChunks()"/>
-        [Pure]
         public IEnumerable<byte[]> EncodeIntoChunks()
         {
             // FIXME: avoid duplication between this and EncodeToStream()
             long length = _dictionaryPrefix.LongLength;
             yield return _dictionaryPrefix;
 
-            var @enum = _dict ?? _pairs ?? Enumerable.Empty<KeyValuePair<IKey, IValue>>();
-            IEnumerable<ValueTuple<byte?, byte[], IValue>> rawPairs =
-                from pair in @enum
-                select (
-                    pair.Key.KeyPrefix,
-                    pair.Key.EncodeAsByteArray(),
-                    pair.Value
-                );
-            IEnumerable<ValueTuple<byte?, byte[], IValue>> orderedPairs = rawPairs.OrderBy(
-                triple => (triple.Item1, triple.Item2),
-                keyPairComparer
-            );
-            var byteArrayComparer = default(ByteArrayComparer);
-            (byte? keyPrefix, byte[] key)? prev = null;
-            foreach ((byte? keyPrefix, byte[] key, IValue value) in orderedPairs)
+            foreach (KeyValuePair<IKey, IValue> pair in this)
             {
-                // Skip duplicates
-                if (_dict is null && prev is { } p)
-                {
-                    if (p.keyPrefix == keyPrefix && byteArrayComparer.Compare(p.key, key) == 0)
-                    {
-                        continue;
-                    }
-                }
-
-                if (keyPrefix != null)
+                if (pair.Key.Type == ValueType.Text)
                 {
                     yield return _unicodeKeyPrefix;
                     length += _unicodeKeyPrefix.LongLength;
                 }
 
+                byte[] key = pair.Key.EncodeAsByteArray();
                 byte[] keyLengthBytes = Encoding.ASCII.GetBytes(
                     key.Length.ToString(CultureInfo.InvariantCulture)
                 );
@@ -513,13 +450,11 @@ namespace Bencodex.Types
                 length += CommonVariables.Separator.LongLength;
                 yield return key;
                 length += key.LongLength;
-                foreach (byte[] chunk in value.EncodeIntoChunks())
+                foreach (byte[] chunk in pair.Value.EncodeIntoChunks())
                 {
                     yield return chunk;
                     length += chunk.LongLength;
                 }
-
-                prev = (keyPrefix, key);
             }
 
             yield return CommonVariables.Suffix;
@@ -528,58 +463,33 @@ namespace Bencodex.Types
         }
 
         /// <inheritdoc cref="IValue.EncodeToStream(Stream)"/>
-        [Pure]
         public void EncodeToStream(Stream stream)
         {
             // FIXME: avoid duplication between this and EncodeIntoChunks()
             long startPos = stream.Position;
             stream.WriteByte(_dictionaryPrefix[0]);
 
-            var @enum = _dict ?? _pairs ?? Enumerable.Empty<KeyValuePair<IKey, IValue>>();
-            IEnumerable<ValueTuple<byte?, byte[], IValue>> rawPairs =
-                from pair in @enum
-                select (
-                    pair.Key.KeyPrefix,
-                    pair.Key.EncodeAsByteArray(),
-                    pair.Value
-                );
-            IEnumerable<ValueTuple<byte?, byte[], IValue>> orderedPairs = rawPairs.OrderBy(
-                triple => (triple.Item1, triple.Item2),
-                keyPairComparer
-            );
-            var byteArrayComparer = default(ByteArrayComparer);
-            (byte? keyPrefix, byte[] key)? prev = null;
-            foreach ((byte? keyPrefix, byte[] key, IValue value) in orderedPairs)
+            foreach (KeyValuePair<IKey, IValue> pair in this)
             {
-                // Skip duplicates
-                if (_dict is null && prev is { } p)
-                {
-                    if (p.keyPrefix == keyPrefix && byteArrayComparer.Compare(p.key, key) == 0)
-                    {
-                        continue;
-                    }
-                }
-
-                if (keyPrefix != null)
+                if (pair.Key.Type == ValueType.Text)
                 {
                     stream.WriteByte(_unicodeKeyPrefix[0]);
                 }
 
+                byte[] key = pair.Key.EncodeAsByteArray();
                 var keyLen =
                     key.Length.ToString(CultureInfo.InvariantCulture);
                 var keyLenBytes = Encoding.ASCII.GetBytes(keyLen);
                 stream.Write(keyLenBytes, 0, keyLenBytes.Length);
                 stream.WriteByte(CommonVariables.Separator[0]);
                 stream.Write(key, 0, key.Length);
-                value.EncodeToStream(stream);
-                prev = (keyPrefix, key);
+                pair.Value.EncodeToStream(stream);
             }
 
             stream.WriteByte(CommonVariables.Suffix[0]);
             _encodingLength = stream.Position - startPos;
         }
 
-        [Pure]
         public override string ToString() =>
             $"{nameof(Bencodex)}.{nameof(Bencodex.Types)}.{nameof(Dictionary)} {Inspection}";
     }