+++ CHANGES.md
@@ -8,6 +8,8 @@ To be released.
 
  -  `Bencodex.Types.List` struct became a class.  [[#51]]
  -  `Bencodex.Types.Dictionary` readonly struct became a class.  [[#51]]
+ -  `Bencodex.Types.Dictionary` now implements
+    `IEquatable<Bencodex.Types.Dictionary>` interface.  [[#51]]
  -  Added `Bencodex.Types.ValueType` enum type.  [[#50]]
  -  Bencodex values now have their unqiue fingerprints:  [[#50]]
      -  Added `Bencodex.Types.Fingerprint` readonly struct.
@@ -25,7 +27,7 @@ To be released.
     constructor.  [[#50]]
  -  Added `List(in ImmutableArray<IValue>)` constructor.  [[#51]]
  -  Added `List(params IValue[])` constructor.  [[#51]]
- -  Added `Dictionary(in ImmutableDictionary<IKey, IValue>)` constructor.
+ -  Added `Dictionary(in ImmutableSortedDictionary<IKey, IValue>)` constructor.
      [[#51]]
  -  Added `Bencodex.Misc.KeyComparer` class.  [[#51]]
  -  `List.Empty` static property became a static readonly field.  [[#50]]